Load the activity and type filter data differently
The previous implementation used the relationship data which is paginated by mu-cl-resources and results in only the first 20 results being returned.

We now use a query with a large page size as a way to load all the data. This should be fine since we don't expect huge amounts of data, but most likely still more than 20.