Fix an issue with file downloads in Chrome
It seems that calling `revokeObjectURL` breaks the ability to download a file from the built-in PDF viewer. The url is released, so it can't be downloaded anymore. Firefox doesn't have this issue. There is no way for us to know when the user downloaded the file, so we can't delay the `revokeObjectURL` call until that happens.

Instead of choosing a very long (hour+) delay, we just don't call `revokeObjectURL` and let the browser handle it when the tabs are closed. This is more correct, at the cost of keeping the files in memory until the browser cleans it up.