Rename helper file, change to role checking
The name now does not follow conventions, but it wouldn't compile
otherwise. It worked at first, but after a merge with unrelated
code chages, the project suddenly wouldn't compile because of this. I
really have no clue as to why the original name is no longer working.
Also changed the role checking on the role itself, instead of on the
representative, because all of a sudden, working with promises wouldn't
work any longer. This was also really weird.