Fix a timing issue with the out of date messages
It seems that the read requests sometimes contain out of date information after a write action. To work around this we add a one second delay after saving, which should be more than enough for the API to be updated again.

The API does provide a mechanism for this use-case, but we opted for a simpler solution for now.